// lib/docx/types.ts
export type DocumentNode =
  | { type: 'paragraph'; children: InlineNode[] }
  | { type: 'math_block'; latex: string }
  | { type: 'heading'; level: 1 | 2 | 3 | 4 | 5 | 6; children: InlineNode[] }
  | { type: 'list'; ordered: boolean; items: DocumentNode[][] }
  | { type: 'blockquote'; children: DocumentNode[] };

export type InlineNode =
  | { type: 'text'; value: string }
  | { type: 'math_inline'; latex: string }
  | { type: 'strong'; children: InlineNode[] }
  | { type: 'emphasis'; children: InlineNode[] }
  | { type: 'link'; url: string; children: InlineNode[] }
  | { type: 'inline_code'; value: string };